pyproject.toml
[project]
name = "watermark-master"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dongwlin"},
]
dependencies = [
    "pyside6>=6.8.0.2",
    "pillow>=11.0.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "LGPL-3.0"}


[tool.pdm]
distribution = false
